Tar writers convert manifests to Shar placeholders themselves
