+++ lhotse/shar/readers/utils.py
@@ -1,58 +1,4 @@
 import os
-from itertools import islice
-from pathlib import Path
-from typing import Optional, Union
-
-from lhotse import Features, Recording
-from lhotse.array import Array, TemporalArray
-from lhotse.cut import Cut
-from lhotse.utils import Pathlike
-
-
-def fill_shar_placeholder(
-    manifest: Union[Cut, Recording, Features, Array, TemporalArray],
-    data: bytes,
-    tarpath: Pathlike,
-    field: Optional[str] = None,
-) -> None:
-    if isinstance(manifest, Cut):
-        assert (
-            field is not None
-        ), "'field' argument must be provided when filling a Shar placeholder in a Cut."
-        manifest = getattr(manifest, field)
-        fill_shar_placeholder(
-            manifest=manifest, field=field, data=data, tarpath=tarpath
-        )
-
-    tarpath = Path(tarpath)
-
-    if isinstance(manifest, Recording):
-        assert (
-            len(manifest.sources) == 1
-        ), "Multiple AudioSources are not supported yet."
-        manifest.sources[0].type = "memory"
-        manifest.sources[0].source = data
-
-    elif isinstance(manifest, (Features, Array)):
-        manifest.storage_key = data
-        if tarpath.suffix == ".llc":
-            manifest.storage_type = "memory_lilcom"
-        elif tarpath.suffix == ".npy":
-            manifest.storage_type = "memory_npy"
-        else:
-            raise RuntimeError(f"Unknown array/tensor format: {tarpath}")
-
-    elif isinstance(manifest, TemporalArray):
-        manifest.array.storage_key = data
-        if tarpath.suffix == ".llc":
-            manifest.array.storage_type = "memory_lilcom"
-        elif tarpath.suffix == ".npy":
-            manifest.array.storage_type = "memory_npy"
-        else:
-            raise RuntimeError(f"Unknown array/tensor format: {tarpath}")
-
-    else:
-        raise RuntimeError(f"Unknown manifest type: {type(manifest).__name__}")
 
 
 def pytorch_worker_info(group=None):
