+++ lhotse/shar/writers/audio.py
@@ -10,6 +10,7 @@ import torchaudio
 from typing_extensions import Literal
 
 from lhotse import Recording
+from lhotse.shar.utils import to_shar_placeholder
 from lhotse.shar.writers.tar import TarWriter
 
 
@@ -88,6 +89,7 @@ class AudioTarWriter:
         self.tar_writer.write(f"{key}.{self.format}", stream)
 
         # Write text manifest afterwards
+        manifest = to_shar_placeholder(manifest)
         json_stream = BytesIO()
         print(
             json.dumps(manifest.to_dict()),
